+++ iblrig/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "7.0.3"
+__version__ = "7.0.4"
 import logging
 
 import colorlog