+++ test_iblrig/fixtures/test_iblrig_params/IBL/tasks/_blank_task/_blank_task.py
@@ -1,0 +1,5 @@
+import logging
+
+log = logging.getLogger("iblrig")
+
+log.info("Blank task, used for testing...")